requirements.txt
# Drishya — No-Reference Video Quality Assessment
# Install: pip install -r requirements.txt

# ── Core scientific stack ──────────────────────────────────
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
opencv-python>=4.8
Pillow>=10.0

# ── Video decoding ─────────────────────────────────────────
decord>=0.6.0

# ── Deep learning (Phase 4 feature extraction) ────────────
torch>=2.0
torchvision>=0.15
timm>=1.0

# ── Phase 5 UI ────────────────────────────────────────────
streamlit>=1.35
plotly>=5.18

# ── Classical NR metrics ──────────────────────────────────
brisque>=0.0.15
piq>=0.8

# ── Data / annotation handling ────────────────────────────
openpyxl>=3.1

# ── ML utilities ──────────────────────────────────────────
tqdm>=4.65
